// generated/src/app/api-models/cap-calc-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapCalcSetup {
  seqCcalcId: number;
  jobId: string;
  requestUser: string;
  requestDatetime: Date;
  requestType: string;
  action: string;
  status: string;
  capModelId: string;
  capMonth: Date;
  glPostMonth: Date;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  template: string;
  inProcess: string;
}